Enable dump single container and movethe code to data_dumper
